// src/app/rooster/interfaces/unit-ui.interface.ts
export interface unitUI {
  gridUnit: number;
  unitFileGrids: number[];
  unitRankGrids: number[];
  fileYPlaces: number[];
  rankXPlaces: number[];
  RaFRest: number[];
  unitRFWidth: number;
  unitRFHeight: number;
  unitWidth: number;
  unitHeight: number;
  unitWidthScss: string;
  unitRFWidthScss: string;
  unitRFHeightScss: string;
  unitHeightScss: string;
}